Add module structure and first usable class
This class manages a file and a directory; it creates /tmp/etc,
sets /etc/motd to a silly string and creates a copy of it in /tmp/etc/motd
to illustrate the array-like syntax for multiple file resources.

Further commits to this repo should follow a similar structure;
the goal is that people trying out the ruby dsl should be able to add
the root of the repo to their `modulepath` setting and use any
of the subdirectories inside it as if they were installed modules.